refactor/#30 : WebClientConfig에 Client 요청 Bean으로 등록하여 구현( 메소드로 구현한 것은 삭제 )
